update CLASSES
add classes

+++ src/main/java/com/sonnguyen/webdochoioto/dao/UserDAO.java
@@ -7,7 +7,7 @@ import com.sonnguyen.webdochoioto.entity.User;
 public interface UserDAO {
 	public void insert(User user);
 	public void update(User user);
-	public void delete(int id);
-	public User getUserById(int id);
+	public void delete(Long id);
+	public User getUserById(Long id);
 	public List<User> getAllUser();
 }
